First stab at code for the emyek-bot (#1)
I took the liberty of creating a .gitignore and master and develop branches on your repo, then decided I should follow the model we used at KeyMe and not change your repo without your permission.

This code, per your instructions, only does stuff a quarter of the time. It sleeps for a random amount of time up to 8 hours (intended to run at the start of the workday, and finish during work hours), grab a random quote, and do something with that quote. I'm hazy about what should be done, and also suspect you (Zach) already have a more complete list of quotes somewhere else.

* First stab at code

* pr feedback